requirements.txt
h5py==2.9.0
ipython==7.2.0
Keras==2.2.4
numpy==1.16.2
opencv-python==4.0.0.21
openpyxl==2.5.12
Pillow==5.4.1
scikit-learn==0.20.3
scikit-image==0.15.0
tensorflow==1.13.1
imutils==0.5.2
enum34==1.1.6